[Chore, Refactor] 유틸 함수 이동
